refactor(result_set): merge duplicate try-except and move Druid tests to test_druid.py
