+++ tests/unit_tests/result_set_test.py
@@ -264,116 +264,3 @@ def test_empty_column_names_do_not_rename_explicit_synthetic_names() -> None:
     df = result_set.to_pandas_df()
     assert list(df.columns) == ["_col_1", "_col_0"]
     assert df.iloc[0].tolist() == [10, 20]
-
-
-# ---------------------------------------------------------------------------
-# DruidEngineSpec column normalization tests
-#
-# pydruid infers column types from the first row value, which causes two
-# related problems:
-#
-#   Case 1 – Mixed IEEE special-float strings and numbers:
-#     Druid cannot represent NaN/Infinity in JSON, so pydruid emits them as
-#     the strings "NaN", "Infinity", or "-Infinity".  When these appear in a
-#     numeric column, pa.array() raises ArrowInvalid on the mixed str/float
-#     list and the column falls back to string serialisation.
-#
-#   Case 2 – None as the first value:
-#     pydruid's get_type(None) returns Type.STRING, so any nullable numeric
-#     column whose first row is null gets labelled STRING in the cursor
-#     description.  pa.array() succeeds (producing float64) but
-#     data_type() used to return STRING because the cursor description won.
-#
-# DruidEngineSpec overrides normalize_column_values and resolve_column_type
-# to handle both cases.  BaseEngineSpec preserves the original behaviour.
-# ---------------------------------------------------------------------------
-
-
-def test_druid_ieee_special_floats_preserved_as_numeric() -> None:
-    """
-    Case 1, DruidEngineSpec: columns that mix IEEE special-float strings with
-    real numbers must keep their numeric type (specials become null).
-    """
-    from superset.db_engine_specs.druid import DruidEngineSpec
-
-    data = [("NaN",), (1.5,), ("Infinity",), (2.3,), ("-Infinity",), (None,)]
-    description = [("metric", "STRING", None, None, None, None, None)]
-    result_set = SupersetResultSet(data, description, DruidEngineSpec)  # type: ignore
-
-    col = result_set.columns[0]
-    assert col["type"] == "FLOAT"
-
-    df = result_set.to_pandas_df()
-    assert pd.isna(df["metric"].iloc[0])  # "NaN" → null
-    assert df["metric"].iloc[1] == 1.5
-    assert pd.isna(df["metric"].iloc[2])  # "Infinity" → null
-    assert df["metric"].iloc[3] == 2.3
-    assert pd.isna(df["metric"].iloc[4])  # "-Infinity" → null
-    assert pd.isna(df["metric"].iloc[5])  # None → null
-
-
-def test_base_spec_ieee_special_floats_stringified() -> None:
-    """
-    Case 1, BaseEngineSpec: without Druid's override, columns with mixed
-    special-float strings and numbers fall through to string serialisation.
-    """
-    data = [("NaN",), (1.5,), ("Infinity",)]
-    description = [("metric", "STRING", None, None, None, None, None)]
-    result_set = SupersetResultSet(data, description, BaseEngineSpec)  # type: ignore
-
-    col = result_set.columns[0]
-    assert col["type"] == "STRING"
-
-    df = result_set.to_pandas_df()
-    assert df["metric"].iloc[0] == "NaN"
-    assert df["metric"].iloc[1] == "1.5"
-    assert df["metric"].iloc[2] == "Infinity"
-
-
-def test_druid_none_first_value_reports_numeric_type() -> None:
-    """
-    Case 2, DruidEngineSpec: when the cursor description says STRING (pydruid's
-    first-row None inference) but PyArrow correctly infers float64, the column
-    must be reported as FLOAT, not STRING.
-    """
-    from superset.db_engine_specs.druid import DruidEngineSpec
-
-    data = [(None,), (1.5,), (2.3,), (None,), (4.7,)]
-    description = [("metric", "STRING", None, None, None, None, None)]
-    result_set = SupersetResultSet(data, description, DruidEngineSpec)  # type: ignore
-
-    col = result_set.columns[0]
-    assert col["type"] == "FLOAT"
-
-    df = result_set.to_pandas_df()
-    assert pd.isna(df["metric"].iloc[0])
-    assert df["metric"].iloc[1] == 1.5
-    assert df["metric"].iloc[4] == 4.7
-
-
-def test_base_spec_none_first_value_reports_string_type() -> None:
-    """
-    Case 2, BaseEngineSpec: the cursor-description STRING type must continue
-    to win over PyArrow's float64 inference for non-Druid engines.
-    """
-    data = [(None,), (1.5,), (2.3,)]
-    description = [("metric", "STRING", None, None, None, None, None)]
-    result_set = SupersetResultSet(data, description, BaseEngineSpec)  # type: ignore
-
-    col = result_set.columns[0]
-    assert col["type"] == "STRING"
-
-
-def test_non_string_cursor_type_unaffected_by_druid_spec() -> None:
-    """
-    Columns with a non-STRING cursor description type must not be affected by
-    DruidEngineSpec's resolve_column_type override.
-    """
-    from superset.db_engine_specs.druid import DruidEngineSpec
-
-    data = [(1,), (2,), (3,)]
-    description = [("count", "INT", None, None, None, None, None)]
-    result_set = SupersetResultSet(data, description, DruidEngineSpec)  # type: ignore
-
-    col = result_set.columns[0]
-    assert col["type"] == "INT"
